Revert "landing page"
