+++ package.json
@@ -34,7 +34,6 @@
     "react-dom": "18.2.0",
     "react-hot-toast": "^2.4.0",
     "react-icons": "^4.8.0",
-    "sharp": "^0.32.1",
     "swiper": "^9.2.3",
     "tailwindcss": "3.3.1",
     "yup": "^1.0.2"
