+++ tailwind.config.js
@@ -7,13 +7,11 @@ module.exports = {
   ],
   theme: {
     extend: {
-      backdropFilter: {
-        'none': 'none',
-        'blur': 'blur(20px)',
+      backgroundImage: {
+        'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
+        'gradient-conic':
+          'conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))',
       },
-      borderRadius: {
-        extraLarge: '5rem'
-      }
     },
   },
   plugins: [
